feat: Add favorite button
Allow the user to set a restaurant as a favorite. When a restaurant is set as favorite the restaurant data will be updated in IDB and an attemp to update the data in the server will be made. If the server update fails, the request will be tried again when the 'online' event is triggered.